Clean address validation in checkout mutations (#17432)
